feat(sketch.js): allow fromFile api to receive a file as buffer OR path
It is very inconvenient when integrating multiple systems working with sketch 'files' to need to
write to the filesystem each time they would like to use sketch-constructor. Adding the ability for
the file to be either a buffer or path solves this issue cleanly, without needing to change the api.